+++ pkg/reconciliation/construct_statefulset.go
@@ -64,8 +64,7 @@ func newStatefulSetForCassandraDatacenter(
 	sts *appsv1.StatefulSet,
 	rackName string,
 	dc *api.CassandraDatacenter,
-	replicaCount int,
-	openShift bool) (*appsv1.StatefulSet, error) {
+	replicaCount int) (*appsv1.StatefulSet, error) {
 
 	replicaCountInt32 := int32(replicaCount)
 
@@ -113,7 +112,7 @@ func newStatefulSetForCassandraDatacenter(
 
 	nsName := newNamespacedNameForStatefulSet(dc, rackName)
 
-	template, err := buildPodTemplateSpec(dc, rack, legacyInternodeMount(dc, sts), openShift)
+	template, err := buildPodTemplateSpec(dc, rack, legacyInternodeMount(dc, sts))
 	if err != nil {
 		return nil, err
 	}